+++ nsh/Cargo.toml
@@ -14,6 +14,7 @@ rust-version = "1.59.0"
 amplify = "3.12.1"
 streampipes = { path = "../streampipes" }
 ioreactor = { path = "../reactor", features = ["popol"] }
+p2pd = { path = "../p2pd" }
 cyphernet = { git = "https://github.com/Internet2-WG/rust-cyphernet", features = ["ed25519"] }
 ntls = { git = "https://github.com/Internet2-WG/ntls" }
 ed25519-compact = { version = "2.0.2", features = ["ed25519", "x25519", "pem", "ct-codecs"] }
