+++ p2pd/src/nxk_tcp.rs
@@ -1,0 +1,183 @@
+//! Noise_XK streams, connections and sessions based on TCP stream
+
+use std::io::{self, Read, Write};
+use std::net;
+use std::os::fd::{AsRawFd, RawFd};
+
+use cyphernet::addr::{LocalNode, PeerAddr, UniversalAddr};
+use cyphernet::crypto::ed25519::Curve25519;
+use ioreactor::{Controller, IoEv, ReactorApi, Resource};
+
+use crate::noise_xk;
+
+pub type NxkAddr = UniversalAddr<PeerAddr<Curve25519, net::SocketAddr>>;
+
+pub enum NxkMethod {
+    Accept(net::TcpStream, net::SocketAddr),
+    Connect(NxkAddr),
+}
+
+pub struct NxkContext {
+    pub method: NxkMethod,
+    pub local_node: LocalNode<Curve25519>,
+}
+
+pub type NxkStream = noise_xk::Stream<Curve25519, net::TcpStream>;
+
+pub struct NxkSession {
+    stream: NxkStream,
+    socket_addr: net::SocketAddr,
+    peer_addr: Option<NxkAddr>,
+    inbound: bool,
+}
+
+impl NxkSession {
+    pub fn accept(
+        tcp_stream: net::TcpStream,
+        remote_socket_addr: net::SocketAddr,
+        local_node: LocalNode<Curve25519>,
+    ) -> Self {
+        Self {
+            stream: NxkStream::upgrade(tcp_stream, local_node),
+            socket_addr: remote_socket_addr,
+            peer_addr: None,
+            inbound: true,
+        }
+    }
+
+    pub fn connect(nsh_addr: NxkAddr, local_node: LocalNode<Curve25519>) -> io::Result<Self> {
+        // TODO: Use socks5
+        let tcp_stream = net::TcpStream::connect(&nsh_addr)?;
+        let remote_key = nsh_addr.as_remote_addr().to_pubkey();
+        Ok(Self {
+            stream: NxkStream::connect(tcp_stream, remote_key, local_node)?,
+            socket_addr: nsh_addr.to_socket_addr(),
+            peer_addr: Some(nsh_addr),
+            inbound: false,
+        })
+    }
+}
+
+impl AsRawFd for NxkSession {
+    fn as_raw_fd(&self) -> RawFd {
+        self.stream.as_raw_fd()
+    }
+}
+
+impl Read for NxkSession {
+    fn read(&mut self, buf: &mut [u8]) -> io::Result<usize> {
+        self.stream.read(buf)
+    }
+}
+
+impl Write for NxkSession {
+    fn write(&mut self, buf: &[u8]) -> io::Result<usize> {
+        self.stream.write(buf)
+    }
+
+    fn flush(&mut self) -> io::Result<()> {
+        self.stream.flush()
+    }
+}
+
+impl Resource for NxkSession {
+    type Id = RawFd;
+    type Context = NxkContext;
+    type Cmd = ();
+    type Error = io::Error;
+
+    fn with(context: Self::Context, controller: Controller<Self>) -> Result<Self, Self::Error>
+    where
+        Self: Sized,
+    {
+        match context.method {
+            NxkMethod::Accept(tcp_stream, remote_socket_addr) => Ok(NxkSession::accept(
+                tcp_stream,
+                remote_socket_addr,
+                context.local_node,
+            )),
+            NxkMethod::Connect(nsh_addr) => NxkSession::connect(nsh_addr, context.local_node),
+        }
+    }
+
+    fn id(&self) -> Self::Id {
+        self.stream.as_raw_fd()
+    }
+
+    fn io_ready(&mut self, io: IoEv) -> Result<(), Self::Error> {
+        if io.is_readable { /* TODO: Do read */ }
+        if io.is_writable {
+            self.flush()?;
+        }
+        Ok(())
+    }
+
+    fn handle_cmd(&mut self, cmd: Self::Cmd) -> Result<(), Self::Error> {
+        // TODO: Do writes, like
+        // self.stream.write_all(&data)?
+        todo!()
+    }
+
+    fn handle_err(&mut self, err: Self::Error) -> Result<(), Self::Error> {
+        todo!()
+    }
+}
+
+pub struct NxkListener<R>
+where
+    R: Resource,
+    R::Context: From<NxkContext>,
+{
+    socket: net::TcpListener,
+    local_node: LocalNode<Curve25519>,
+    controller: Controller<R>,
+}
+
+impl<R> Resource<R> for NxkListener<R>
+where
+    R: Resource,
+    R::Context: From<NxkContext>,
+{
+    type Id = RawFd;
+    type Context = (LocalNode<Curve25519>, net::SocketAddr);
+    type Cmd = ();
+    type Error = io::Error;
+
+    fn with(context: Self::Context, controller: Controller<R>) -> Result<Self, Self::Error>
+    where
+        Self: Sized,
+    {
+        let socket = net::TcpListener::bind(context.1)?;
+        Ok(Self {
+            socket,
+            controller,
+            local_node: context.0,
+        })
+    }
+
+    fn id(&self) -> Self::Id {
+        self.socket.as_raw_fd()
+    }
+
+    fn io_ready(&mut self, _: IoEv) -> Result<(), Self::Error> {
+        let (stream, peer_socket_addr) = self.socket.accept()?;
+        let nsh_info = NxkContext {
+            method: NxkMethod::Accept(stream, peer_socket_addr),
+            local_node: self.local_node.clone(),
+        };
+        self.controller
+            .connect(nsh_info.into())
+            .map_err(|_| io::ErrorKind::NotConnected)?;
+        Ok(())
+    }
+
+    fn handle_cmd(&mut self, cmd: Self::Cmd) -> Result<(), Self::Error> {
+        // Listener does not support any commands
+        Ok(())
+    }
+
+    fn handle_err(&mut self, err: Self::Error) -> Result<(), Self::Error> {
+        // Listener does not know how to handle errors, so it just propagates them
+        Err(err)
+    }
+}
